refactor(api): remove username from client initialization
This change removes the `username` parameter from the Kimai client
initialization and related API calls. The authentication now relies
solely on the API token, this is the new default for the Kimai API.

BREAKING CHANGE: The `NewKimaiClient` function no longer accepts a
`username` parameter. Update all client initializations to exclude
the username and use only the API token for authentication.

Signed-off-by: Anne Douwe Bouma <ad.bouma@247kooi.com>